fix(http): guard against more null users in controllers

+++ lib/Controller/SmimeCertificatesController.php
@@ -44,6 +44,10 @@ class SmimeCertificatesController extends Controller {
 	 */
 	#[TrapError]
 	public function index(): JsonResponse {
+		if ($this->userId === null) {
+			return JsonResponse::fail([], Http::STATUS_UNAUTHORIZED);
+		}
+
 		$certificates = $this->certificateService->findAllCertificates($this->userId);
 		$certificates = array_map(fn (SmimeCertificate $certificate) => $this->certificateService->enrichCertificate($certificate), $certificates);
 		return JsonResponse::success($certificates);
@@ -59,6 +63,10 @@ class SmimeCertificatesController extends Controller {
 	 */
 	#[TrapError]
 	public function destroy(int $id): JsonResponse {
+		if ($this->userId === null) {
+			return JsonResponse::fail([], Http::STATUS_UNAUTHORIZED);
+		}
+
 		$this->certificateService->deleteCertificate($id, $this->userId);
 		return JsonResponse::success();
 	}
@@ -73,6 +81,10 @@ class SmimeCertificatesController extends Controller {
 	 */
 	#[TrapError]
 	public function create(): JsonResponse {
+		if ($this->userId === null) {
+			return JsonResponse::fail([], Http::STATUS_UNAUTHORIZED);
+		}
+
 		// TODO: What about PKCS12 certificates?
 		// They need to be decrypted by the client because they are protected by a password.
 		// We could use
